pyproject.toml
[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "eigsep_observing"
version = "1.0.0"
description = "Control code for EIGSEP observing."
readme = "README.md"
license = {text = "MIT"}
authors = [{ name="Christian Hellum Bye", email="cbh@berkeley.edu" }]
requires-python = ">=3.10"
classifiers = [
  "Intended Audience :: Science/Research",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Topic :: Scientific/Engineering :: Astronomy",
]
dependencies = [
  "h5py",
  "numpy",
  "pyyaml",
  "flask",
  "plotly",
  "picohost~=3.0",
  "eigsep-vna~=1.3",
  "eigsep_redis~=2.1",
]

[project.optional-dependencies]
dev = [
  "build",
  "twine",
  "ruff",
  "fakeredis",
  "pyserial-mock",
  "pytest",
  "pytest-cov",
  "pytest-timeout",
  "pytest-xdist",
]
interactive = [
  "matplotlib",
  "ipython",
]

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"eigsep_observing" = [
  "config/*.yaml",
  "data/*.fpg",
  "live_status/templates/*.html",
  "live_status/static/css/*.css",
  "live_status/static/js/*.js",
]

[tool.pytest.ini_options]
testpaths = ["tests", "src/eigsep_observing/contract_tests"]
pythonpath = ["src"]
addopts = [
  "--cov=eigsep_observing",
  "--cov-report=term-missing",
  "--cov-report=xml",
  "--junitxml=junit.xml",
  "-o junit_family=legacy",
  "--timeout=60",
]

[tool.coverage.run]
omit = [
  "src/eigsep_observing/contract_tests/*",
  "src/eigsep_observing/_test_fixtures.py",
]

[tool.ruff]
line-length = 79

[tool.ruff.lint]
extend-ignore = ["E203"]

[tool.ruff.lint.per-file-ignores]
"src/*/__init__.py" = ["F401"]
"scripts/*.py" = ["E402"]
